Refactor C Skeleton Generation
Decouple the data parsing from the UI
